+++ tests/unit/interface/no_tango/test_cm_unhappy_paths.py
@@ -74,7 +74,7 @@ def test_periodic_noise_diode_pars_with_invalid_states(
         " Device is not in STANDBY state."
         " Current state: UNKNOWN",
     ):
-        component_manager.set_periodic_noise_diode_pars([1.0, 2.0, 3.0])
+        component_manager.set_periodic_noise_diode_pars([1, 2, 3])
 
 
 @pytest.mark.unit
@@ -109,4 +109,4 @@ def test_pseudo_random_noise_diode_pars_with_invalid_states(
         " Device is not in STANDBY state."
         " Current state: UNKNOWN",
     ):
-        component_manager.set_pseudo_random_noise_diode_pars([1.0, 2.0, 3.0])
+        component_manager.set_pseudo_random_noise_diode_pars([1, 2, 3])