+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 ## unreleased
 *************
+- Updated periodicNoiseDiodePars and pseudoRandomNoiseDiodePars to be DevVarLong64Array
 
 Version 7.4.0
 *************